refactor the login implementation
