+++ .travis.yml
@@ -1,6 +1,10 @@
 language: node_js
 node_js:
   - "8"
+before_script:
+  - psql -c 'create database EduGateTestDB;' -U postgres
+  - npm db-init-test
+  
 before_install:
   - npm i -g npm@6.4.1
