+++ server/config/config.js
@@ -15,6 +15,7 @@ const config = {
     username: process.env.DB_USERNAME,
     password: process.env.DB_PASSWORD,
     database: process.env.DB_TEST_NAME,
+    logging: true,
     host: process.env.DB_HOST,
     port: process.env.DB_PORT,
     dialect: 'postgres'
