+++ server/controllers/users/login.js
@@ -1,5 +1,5 @@
 import bcrypt from 'bcryptjs';
-import db from '../../server/models/index'
+import db from '../../models'
 
 let { Users } = db;
 
@@ -11,15 +11,17 @@ login.get = (req, res) => {
 
 login.post = (req, res) => {
   const { username, password } = req.body;
-  Users.findAll({
+  Users.findOne({
     attributes: ['password', 'user_uid', 'role'],
     where: {
         username
     }
   })
   .then(user => {
-    if (!user || !user.length) return res.send({err: "No User Found"})
-    const userData = user[0].dataValues;
+    if (!user) {
+      return res.status(404).json({err: "No User Found"})
+    }
+    const userData = user.dataValues;
     bcrypt.compare(password, userData.password)
     .then(match => {
       if (match) {
@@ -27,13 +29,19 @@ login.post = (req, res) => {
           user_uid: userData.user_uid,
           role: userData.role,
         }
-        req.session = user
+        req.session = user;
+        res.status(200).json({
+          message: "User login successfully"
+        })
       }
-      res.send({match})
+      return res.status(404).json({
+        err: "Password is incorrect. Please try again"
+      })
     })
-    .catch(err => console.log(err))
   })
-  .catch(err => console.log(err))
+  .catch(err => res.status(500).json({
+      err
+  }))
 };
 
 export { login }
